requirements.txt
Flask-InfluxDB~=0.3.2
Flask~=1.1.2
pyOpenSSL~=19.1.0
python-aqi~=0.6.1
threema.gateway~=3.1.0
